Split test_proxy, rename others according to modules
