+++ tests/client/conftest.py
@@ -22,6 +22,9 @@ TEST_DASH_ENDPOINT = os.getenv('DASH_ENDPOINT', 'http://33.33.33.51:8080/api/')
 TEST_HS_ENDPOINT = os.getenv('HS_ENDPOINT',
                              'http://storage.vm.scrapinghub.com')
 
+# use some fixed timestamp to represent current time
+TEST_TS = 1476803148638
+
 # vcrpy creates the cassetes automatically under VCR_CASSETES_DIR
 VCR_CASSETES_DIR = 'tests/client/cassetes'
