--- tests/client/test_proxy.py
@@ -1,306 +1,0 @@
-import json
-import types
-from numbers import Integral
-
-import pytest
-from six.moves import range
-
-from scrapinghub.client.utils import LogLevel
-from scrapinghub.hubstorage.serialization import mpdecode
-
-from .conftest import TEST_PROJECT_ID
-
-# use some fixed timestamp to represent current time
-TEST_TS = 1476803148638
-
-
-def _add_test_logs(job):
-    job.logs.log('simple-msg1')
-    job.logs.log('simple-msg2', ts=TEST_TS)
-    job.logs.log('simple-msg3', level=LogLevel.DEBUG)
-
-    job.logs.debug('debug-msg')
-    job.logs.info('info-msg')
-    job.logs.warning('warning-msg')
-    job.logs.warn('warn-msg')
-    job.logs.error('error-msg')
-
-    # test raw write interface
-    job.logs.write({'message': 'test',
-                    'level': LogLevel.INFO,
-                    'time': TEST_TS})
-    job.logs.flush()
-
-
-def test_logs_base(spider):
-    job = spider.jobs.schedule()
-    assert list(job.logs.iter()) == []
-    assert job.logs.batch_write_start() == 0
-    _add_test_logs(job)
-    log1 = job.logs.get(0)
-    assert log1['level'] == 20
-    assert log1['message'] == 'simple-msg1'
-    assert isinstance(log1['time'], Integral) and log1['time'] > 0
-    assert job.logs.stats() == {
-        'counts': {'10': 2, '20': 4, '30': 2, '40': 1},
-        'totals': {'input_bytes': 91, 'input_values': 9}
-    }
-    job.logs.close()
-
-
-def test_logs_iter(spider):
-    job = spider.jobs.schedule()
-    _add_test_logs(job)
-
-    logs1 = job.logs.iter()
-    assert isinstance(logs1, types.GeneratorType)
-    assert next(logs1).get('message') == 'simple-msg1'
-    assert next(logs1).get('message') == 'simple-msg2'
-    assert next(logs1).get('message') == 'simple-msg3'
-
-    # testing offset
-    logs2 = job.logs.iter(offset=3)
-    assert next(logs2).get('message') == 'debug-msg'
-    assert next(logs2).get('message') == 'info-msg'
-
-    # testing level
-    logs3 = job.logs.iter(level='ERROR')
-    assert next(logs3).get('message') == 'error-msg'
-    with pytest.raises(StopIteration):
-        next(logs3).get('message')
-
-
-def test_logs_list(spider):
-    job = spider.jobs.schedule()
-    _add_test_logs(job)
-
-    logs1 = job.logs.list()
-    assert isinstance(logs1, list)
-    assert len(logs1) == 9
-    assert logs1[0].get('message') == 'simple-msg1'
-    assert logs1[1].get('message') == 'simple-msg2'
-    assert logs1[2].get('message') == 'simple-msg3'
-
-    # testing offset
-    logs2 = job.logs.list(offset=3)
-    assert len(logs2) == 6
-    assert logs2[0].get('message') == 'debug-msg'
-    assert logs2[1].get('message') == 'info-msg'
-
-    # testing level
-    logs3 = job.logs.list(level='ERROR')
-    assert len(logs3) == 1
-    assert logs3[0].get('message') == 'error-msg'
-
-
-def test_logs_list_filter(spider):
-    job = spider.jobs.schedule()
-    _add_test_logs(job)
-
-    logs1 = job.logs.list(filter='["message", "contains", ["simple"]]')
-    assert isinstance(logs1, list)
-    assert len(logs1) == 3
-    assert logs1[0].get('message') == 'simple-msg1'
-
-    logs2 = job.logs.list(filter=[['message', 'contains', ['simple']]])
-    assert len(logs2) == 3
-
-    logs3 = job.logs.list(filter=[('message', 'contains', ['simple'])])
-    assert len(logs3) == 3
-
-
-def test_logs_iter_raw_json(spider):
-    job = spider.jobs.schedule()
-    _add_test_logs(job)
-
-    logs0 = job.logs.iter_raw_json(offset=2)
-    raw_log0 = next(logs0)
-    log0 = json.loads(raw_log0)
-    assert log0.get('message') == 'simple-msg3'
-    assert log0.get('_key')
-    assert isinstance(log0.get('time'), Integral)
-    assert log0.get('level') == 10
-
-    logs1 = job.logs.iter_raw_json(level='ERROR')
-    raw_log1 = next(logs1)
-    log1 = json.loads(raw_log1)
-    assert log1.get('message') == 'error-msg'
-
-
-def test_logs_iter_raw_msgpack(spider):
-    job = spider.jobs.schedule()
-    _add_test_logs(job)
-
-    logs1 = job.logs.iter_raw_msgpack(offset=2)
-    assert isinstance(logs1, types.GeneratorType)
-    unpacked_logs1 = list(mpdecode(logs1))
-    assert unpacked_logs1[0].get('message') == 'simple-msg3'
-
-    logs2 = job.logs.iter_raw_msgpack(level='ERROR')
-    unpacked_logs2 = list(mpdecode(logs2))
-    assert unpacked_logs2[0].get('message') == 'error-msg'
-
-
-def _add_test_requests(job):
-    r1 = job.requests.add(
-        url='http://test.com/', status=200, method='GET',
-        rs=1337, duration=5, parent=None, ts=TEST_TS)
-    job.requests.add(
-        url='http://test.com/2', status=400, method='POST',
-        rs=0, duration=1, parent=r1, ts=TEST_TS + 1)
-    job.requests.add(
-        url='http://test.com/3', status=400, method='PUT',
-        rs=0, duration=1, parent=r1, ts=TEST_TS + 2, fp='1234')
-    job.requests.flush()
-
-
-def test_requests_iter(spider):
-    job = spider.jobs.schedule(meta={'state': 'running'})
-    _add_test_requests(job)
-    job.requests.close()
-    rr = job.requests.iter()
-    assert next(rr) == {
-        'url': 'http://test.com/', 'status': 200, 'rs': 1337,
-        'time': TEST_TS, 'duration': 5, 'method': 'GET',
-    }
-    assert next(rr) == {
-        'url': 'http://test.com/2', 'status': 400, 'rs': 0,
-        'time': TEST_TS + 1, 'duration': 1, 'method': 'POST',
-        'parent': 0,
-    }
-    assert next(rr) == {
-        'url': 'http://test.com/3', 'status': 400, 'rs': 0,
-        'time': TEST_TS + 2, 'method': 'PUT', 'fp': '1234',
-        'parent': 0, 'duration': 1,
-    }
-    with pytest.raises(StopIteration):
-        next(rr)
-
-
-def test_requests_iter_raw_json(spider):
-    job = spider.jobs.schedule()
-    _add_test_requests(job)
-    job.requests.close()
-
-    rr = job.requests.iter_raw_json()
-    raw_req = next(rr)
-    req = json.loads(raw_req)
-    assert req.get('url') == 'http://test.com/'
-    assert req.get('status') == 200
-    next(rr), next(rr)
-    with pytest.raises(StopIteration):
-        next(rr)
-
-
-def _add_test_samples(job):
-    job.samples.write([TEST_TS, 1, 2, 3])
-    job.samples.write([TEST_TS + 1, 5, 9, 4])
-    job.samples.flush()
-    job.samples.close()
-
-
-def test_samples_iter(spider):
-    job = spider.jobs.schedule(meta={'state': 'running'})
-    assert list(job.samples.iter()) == []
-    _add_test_samples(job)
-
-    o = job.samples.iter()
-    assert next(o) == [TEST_TS, 1, 2, 3]
-    assert next(o) == [TEST_TS + 1, 5, 9, 4]
-    with pytest.raises(StopIteration):
-        next(o)
-
-
-def test_samples_list(spider):
-    job = spider.jobs.schedule(meta={'state': 'running'})
-    _add_test_samples(job)
-    o = job.samples.list()
-    assert isinstance(o, list)
-    assert len(o) == 2
-    assert o[0] == [TEST_TS, 1, 2, 3]
-    assert o[1] == [TEST_TS + 1, 5, 9, 4]
-
-
-def _add_test_items(job):
-    for i in range(3):
-        job.items.write({'id': i, 'data': 'data' + str(i)})
-    job.items.flush()
-    job.items.close()
-
-
-def test_items_iter(spider):
-    job = spider.jobs.schedule(meta={'state': 'running'})
-    _add_test_items(job)
-
-    o = job.items.iter()
-    assert next(o) == {'id': 0, 'data': 'data0'}
-    assert next(o) == {'id': 1, 'data': 'data1'}
-    next(o)
-    with pytest.raises(StopIteration):
-        next(o)
-
-    o = job.items.iter(offset=2)
-    assert next(o) == {'id': 2, 'data': 'data2'}
-    with pytest.raises(StopIteration):
-        next(o)
-
-    o = job.items.iter_raw_json(offset=2)
-    item = json.loads(next(o))
-    assert item['id'] == 2
-    assert item['data'] == 'data2'
-    with pytest.raises(StopIteration):
-        next(o)
-
-    msgpacked_o = job.items.iter_raw_msgpack(offset=2)
-    o = mpdecode(msgpacked_o)
-    assert item['id'] == 2
-    assert item['data'] == 'data2'
-
-
-def test_items_list(spider):
-    job = spider.jobs.schedule(meta={'state': 'running'})
-    _add_test_items(job)
-
-    o = job.items.list()
-    assert isinstance(o, list)
-    assert len(o) == 3
-    assert o[0] == {'id': 0, 'data': 'data0'}
-    assert o[1] == {'id': 1, 'data': 'data1'}
-    assert o[2] == {'id': 2, 'data': 'data2'}
-
-
-def _add_test_activity(project):
-    activity = project.activity
-    jobkey = TEST_PROJECT_ID + '/2/3'
-    events = [{'event': 'job:completed', 'job': jobkey, 'user': 'jobrunner'},
-              {'event': 'job:cancelled', 'job': jobkey, 'user': 'john'}]
-    activity.add(events)
-
-
-def test_activity_wrong_project(project):
-    event = {'event': 'job:completed', 'job': '123/1/1', 'user': 'user'}
-    with pytest.raises(ValueError):
-        project.activity.add(event)
-
-
-def test_activity_iter(project):
-    _add_test_activity(project)
-    activity = project.activity.iter()
-    assert isinstance(activity, types.GeneratorType)
-    activity_item = next(activity)
-    assert activity_item == {'event': 'job:cancelled',
-                             'job': TEST_PROJECT_ID + '/2/3',
-                             'user': 'john'}
-
-
-def test_activity_list(project):
-    _add_test_activity(project)
-    activity = project.activity.list(count=2)
-    assert isinstance(activity, list)
-    assert len(activity) == 2
-    assert activity[0] == {'event': 'job:cancelled',
-                           'job': TEST_PROJECT_ID + '/2/3',
-                           'user': 'john'}
-    assert activity[1] == {'event': 'job:completed',
-                           'job': TEST_PROJECT_ID + '/2/3',
-                           'user': 'jobrunner'}
