+++ tests/client/test_spiders.py
